Keep menu bar icon consistent regardless of errors
The menu bar icon now stays as the active icon at all times, instead of
changing to a warning triangle when grammar errors are detected. This
provides a more consistent visual experience in the menu bar.

- Changed AnalysisCoordinator to always set icon state to .active
- Removed .error icon state change when floating indicator is shown
- Added comment explaining user preference for consistent icon

Related to user feedback: menu bar icon should not change based on error state

Signed-off-by: Philip Schmid <philip@textwarden.io>